Write into txn and txn_participation tables in parallel with other import procedures.
